% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{averaged_dataset}
\alias{averaged_dataset}
\title{Builds an averaged dataset from raw data}
\usage{
averaged_dataset(data, na.omit = T)
}
\arguments{
\item{data}{A data frame containing raw source and mixture data.}

\item{na.omit}{Boolean to omit or not NA values when computing the mean and SD}
}
\value{
A data frame representing the averaged dataset.
}
\description{
Generates an averaged dataset from individual (non-averaged) observations.
}
