% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot.R
\name{boxPlot}
\alias{boxPlot}
\title{Box and whiskers plot}
\usage{
boxPlot(data, columns = 1:ncol(data) - 2, ncol = 3)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{columns}{Numeric vector containing the index of the columns in the chart (the first column refers to the first variable)}

\item{ncol}{Number of charts per row}
}
\description{
The boxplot compactly shows the distribution of a continuous variable. It displays five summary statistics (the median, two hinges and two whiskers), and all "outlying" points individually.
}
