% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_database}
\alias{check_database}
\title{Verify the integrity of a sediment unmixing database}
\usage{
check_database(data)
}
\arguments{
\item{data}{A data frame to be checked.}
}
\value{
A logical value (`TRUE` if the database is valid, `FALSE` otherwise).
  If the check fails, the function will also print a descriptive error message.
}
\description{
Verify the integrity of a sediment unmixing database
}
\keyword{internal}
