% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_Method.R
\name{CI_Method}
\alias{CI_Method}
\title{Conservativeness Index (CI)}
\usage{
CI_Method(data, points = 2000, Means = F, seed = 123456L)
}
\arguments{
\item{data}{Data frame containing sediment source and mixtures}

\item{points}{Number of solutions per tracer}

\item{Means}{Boolean to switch when using mean and sd data}

\item{seed}{Seed for the random number generator}
}
\value{
List of data frames containing the selected number of possible prediction for each tracer. The last lists correspond to the dataset and the CI index.
}
\description{
The function quantify the predictions of each individual tracer in your dataset and calculate the CI value of each tracer.
}
