% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Correlation matrix chart}
\usage{
correlationPlot(data, columns = c(1:ncol(data) - 1), mixtures = F)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{columns}{Numeric vector containing the index of the columns in the chart (the first column refers to the grouping variable)}

\item{mixtures}{Boolean to include or exclude the mixture samples in the chart}
}
\description{
The function displays a correlation matrix of each of the properties divided by the different sources to help the user in the decision.
}
