% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_4s.R
\name{cr_4s}
\alias{cr_4s}
\title{CR values of each sediment mixture for four sources}
\usage{
cr_4s(source, mixture, maxiter = 2000, seed = 123456)
}
\arguments{
\item{source}{Data frame containing the sediment sources from a dataset}

\item{mixture}{Data frame containing one of the dataset mixtures}

\item{maxiter}{Number of iteration for each tracer}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the CR score for each tracer.
}
\description{
Compute the consensus ranking method for the selected dataset.
}
