% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_3s.R
\name{cts_3s}
\alias{cts_3s}
\title{CTS method for 3 sources}
\usage{
cts_3s(source, mixture, sol)
}
\arguments{
\item{source}{Data frame containing the sediment sources from a dataset}

\item{mixture}{Data frame containing one of the dataset mixtures}

\item{sol}{Selected solution from the pairs/triplet functions}
}
\value{
Data frame containing the CR score for each tracer.
}
\description{
Compute the CTS for the selected pairs or triplets of tracers.
}
