% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_select_5s.R
\name{CTS_error_5s}
\alias{CTS_error_5s}
\title{Evaluate the mathematical consistency of a tracer selection for an apportionment solution.}
\usage{
CTS_error_5s(source, mixture, solution)
}
\arguments{
\item{source}{Data frame containing the sediment sources from a dataset.}

\item{mixture}{Data frame containing one of the dataset mixtures.}

\item{solution}{A vector containing the apportionment result.}
}
\value{
Data frame containing the normalized error for each tracer.
}
\description{
This function assesses the mathematical consistency of a tracer selection for an apportionment result by computing the normalized error between the predicted and observed tracer concentrations in the virtual mixture. A low normalized error for all tracers indicates a consistent tracer selection. This function can be used to diagnose problems in the results of fingerprinting models and also to extend a minimal tracer combination obtained from the `cts_error` function ensuring its mathematical consistency.
}
\keyword{internal}
