% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_explore_quartets.R
\name{CTS_seeds_quartets}
\alias{CTS_seeds_quartets}
\title{Extract all possible tracer combinations with four tracers.}
\usage{
CTS_seeds_quartets(source, mixture, iter = 1000, rng_init = NULL)
}
\arguments{
\item{source}{Data frame containing the sediment sources from a dataset.}

\item{mixture}{Data frame containing one of the dataset mixtures.}

\item{iter}{Iterations in the variability analysis of each tracer combination.}

\item{rng_init}{An integer value used to initialize the random number generator (RNG). Providing a starting value ensures that the sequence of random numbers generated is reproducible. This is useful for debugging, testing, and comparing results across different runs. If no value is provided, a random one will be generated.}
}
\value{
A data frame containing all possible tracer combinations from the dataset. Each combination is characterized by its corresponding average solution and dispersion (standard deviation), as well as the percentage of solutions that fall within the physically feasible space.
}
\description{
This function generates a list of all possible tracer combinations to identify the most discriminant and serves as a seed to build a consistent tracer selection in a subsequent step. This analysis explores minimal tracer combinations (four tracers for five sources) and solves the resulting determined system of equations to assess the variability of each combination. The dispersion of the solution reflects the discriminant capacity of each tracer combination: a lower dispersion indicates a higher discriminant capacity. Typically, the most discriminant tracer combination corresponds to the result of DFA analysis. In this analysis, the solutions are not restricted to the physically feasible space, which can be valuable for identifying problematic tracer selections that might be masked when using constrained unmixing models.
}
\keyword{internal}
