% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_select.R
\name{CTS_select}
\alias{CTS_select}
\title{Extend minimal tracer sets by evaluating mathematical consistency}
\usage{
CTS_select(data, tracers_seeds, seed_id, error_threshold = 0.05)
}
\arguments{
\item{data}{A data frame containing the characteristics of sediment sources and mixtures.}

\item{tracers_seeds}{A data frame containing the output from the `CTS_explore` function.}

\item{seed_id}{A numeric ID to select a specific row from `tracers_seeds`.}

\item{error_threshold}{A numeric value (e.g., 0.05). Only tracers with a normalized error below this value will be retained.}
}
\value{
A data frame containing the normalized error for each tracer.
}
\description{
This function extends a minimal tracer combination obtained from the `CTS_explore` function ensuring its mathematical consistency in order to select optimum tracers to perform the unmix.
}
\details{
The function calculates a normalized error for each tracer to assess the consistency of a given apportionment solution. The method involves first computing a "virtual mixture" by using the proposed apportionment values to perform a weighted average of the source tracer concentrations. The error for each tracer is then the difference between the tracer concentration in the real mixture and the virtual mixture. This error is normalized by the range of the tracer, which is estimated from the extremes of the sources' confidence intervals.

A low normalized error for all tracers (i.e., less than a predefined threshold like $0.05$) indicates a mathematically consistent tracer selection. If most tracers show low errors while a few have high errors, it suggests that those tracers may be non-conservative or less influential on the model's result. Conversely, high normalized errors in most tracers indicate mathematical inconsistency and can point to the existence of multiple partial solutions in the dataset.
}
\references{
Latorre, B., Lizaga, I., Gaspar, L., & Navas, A. (2021). A novel method for analysing
consistency and unravelling multiple solutions in sediment fingerprinting.
*Science of The Total Environment*, *789*, 147804.
}
