% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA_test.R
\name{DFA_test}
\alias{DFA_test}
\title{Discriminant Function Analysis (DFA) Test}
\usage{
DFA_test(data, niveau = 0.1)
}
\arguments{
\item{data}{A data frame containing the characteristics of sediment sources and mixtures.}

\item{niveau}{A numeric value specifying the significance level for the approximate F-test decision.}
}
\value{
A data frame containing only the tracers that pass the DFA test.
}
\description{
Performs a stepwise forward variable selection using the Wilk's Lambda criterion to identify the most discriminant tracers in a dataset.
}
