% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputMixture.R
\name{inputMixture}
\alias{inputMixture}
\title{Input sediment mixtures}
\usage{
inputMixture(data)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}
}
\description{
The function select and extract the sediment mixtures of the raw dataset.
}
