% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputSample.R
\name{inputSample}
\alias{inputSample}
\title{Input sediment mixtures}
\usage{
inputSample(data)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}
}
\description{
The function select and extract the sediment mixtures of the dataset.
}
