% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputSource.R
\name{inputSource}
\alias{inputSource}
\title{Input sediment sources}
\usage{
inputSource(data, na.omit = T)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{na.omit}{Boolean to omit or not NA values when computing the mean and SD}
}
\description{
The function select and extract the source samples of the dataset.
}
