% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_averaged}
\alias{is_averaged}
\title{Check if data is averaged}
\usage{
is_averaged(data)
}
\arguments{
\item{data}{A data frame containing sediment source and mixture data. It is expected to
have columns for tracer data.}
}
\value{
A logical value. Returns `TRUE` if the data frame is formatted for averaged 
data (i.e., contains a column named "n" and a balanced set of "mean_" and "sd_" 
tracer columns). Returns `FALSE` otherwise.
}
\description{
Checks a data frame to determine if it is formatted for averaged data. 
This is determined by verifying the presence of an "n" column and an equal number of columns 
prefixed with "mean_" and "sd_" that correspond to the same tracers.
}
\keyword{internal}
