% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW_test.R
\name{KW_test}
\alias{KW_test}
\title{Kruskal-Wallis rank sum test}
\usage{
KW_test(data, pvalue = 0.05)
}
\arguments{
\item{data}{Data frame containing source and mixtures}

\item{pvalue}{p-value threshold}
}
\value{
Data frame only containing the variables that pass the Kruskal-Wallis test
}
\description{
This function excludes from the original data frame the properties which do not show significant differences between sources.
}
