% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KWTest.R
\name{KWTest}
\alias{KWTest}
\title{Kruskal-Wallis rank sum test}
\usage{
KWTest(data, pvalue = 0.05)
}
\arguments{
\item{data}{Data frame containing source and mixtures}

\item{pvalue}{p-value threshold}
}
\value{
Data frame only containing the variables that pass the Kruskal-Wallis test
}
\description{
This function excludes from the original data frame the properties which do not show significant differences between sources.
}
\examples{
#Load the dataset called "catchment" 
data <- catchment
#Some of the tracer properties are collinear
data <- KWTest(data)
#The test discard those tracer properties that were unable to discriminate between selected sources
}
