% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plot.R
\name{LDA_plot}
\alias{LDA_plot}
\title{Perform and visualize Linear Discriminant Analysis (LDA)}
\usage{
LDA_plot(data, text = TRUE, colors = NULL)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{text}{Boolean to show or not the identification number of each sample point in the plot}

\item{colors}{Allows choosing between a different set of colors in the plots}
}
\description{
The function performs a linear discriminant analysis and displays the data in the relevant dimensions.
}
