% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPlot.R
\name{LDAPlot}
\alias{LDAPlot}
\title{Linear discriminat analysis chart}
\usage{
LDAPlot(data, P3D = FALSE)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{P3D}{Boolean to switch between 2 to 3 dimensional chart}
}
\description{
The function performs a linear discriminant analysis and displays the data in the relevant dimensions.
}
