% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{least_squares}
\alias{least_squares}
\title{least square unmixing}
\usage{
least_squares(data, iter = 1L, seed = 123456L)
}
\arguments{
\item{data}{Data frame containing sediment source and mixtures}

\item{iter}{Iterations in the source variability analysis}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Asses the relative contribution of the potential sediment sources for each sediment mixture in the dataset.
}
