% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{least_squares_c}
\alias{least_squares_c}
\title{least_squares_c}
\usage{
least_squares_c(sources, mixtures, iter = 100L, rng_init = 123456L)
}
\arguments{
\item{sources}{Data frame containing sediment source samples}

\item{mixtures}{Data frame containing mixture samples}

\item{iter}{Iterations in the variability analysis.}

\item{rng_init}{An integer value used to initialize the random number generator (RNG).}
}
\value{
Data frame containing the relative contribution solved by the least squares method
}
\description{
least_squares_c
}
\keyword{internal}
