% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sources.R
\name{merge_sources}
\alias{merge_sources}
\title{merge_sources}
\usage{
merge_sources(data, S0, S1)
}
\arguments{
\item{data}{Data frame containing source and mixtures}

\item{S0}{First source to be merged}

\item{S1}{Second source to be merged}
}
\value{
Data frame containing sediment sources and mixtures
}
\description{
Function that merge two sources into one.
}
