% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Principal component analysis chart}
\usage{
PCA_plot(data, components = c(1, 2), colors = NULL)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{components}{Numeric vector containing the index of the two principal components in the chart}

\item{colors}{Vector of colors to use for the groups in the plot}
}
\description{
The function performs a principal components analysis on the given data matrix and displays a biplot using vqv.ggbiplot package of the results for each different source to help the user in the decision.
}
