% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tracer.R
\name{remove_tracer}
\alias{remove_tracer}
\title{remove_tracer}
\usage{
remove_tracer(data, rem.traz)
}
\arguments{
\item{data}{Data frame containing source and mixtures}

\item{rem.traz}{Tracers to be removed}
}
\value{
Data frame containing sediment sources and mixtures
}
\description{
Function that excludes the properties of the sediment mixture/s selected by the user.
}
