% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triangles_random_c}
\alias{triangles_random_c}
\title{triangles_random_c}
\usage{
triangles_random_c(
  sources,
  mixtures,
  tracer = 0L,
  iter = 100L,
  rng_init = 123456L
)
}
\arguments{
\item{sources}{Data frame containing sediment source samples}

\item{mixtures}{Data frame containing mixture samples}

\item{tracer}{Tracer in which implement the function}

\item{iter}{Iterations in the variability analysis.}

\item{rng_init}{An integer value used to initialize the random number generator (RNG).}
}
\value{
List of data frames containing all the possible prediction for each tracer
}
\description{
triangles_random_c
}
\keyword{internal}
