% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unmix_c}
\alias{unmix_c}
\title{Unmix sediment mixtures using constrained optimization and classical variability propagation}
\usage{
unmix_c(
  sources,
  mixtures,
  variability,
  iter = 100L,
  resolution = 100L,
  rng_init = 123456L
)
}
\arguments{
\item{sources}{Data frame containing sediment source samples}

\item{mixtures}{Data frame containing mixture samples}

\item{variability}{Integer specifying the type of variability to calculate.
Possible values are 0 for Standard Deviation or 1 for Standard Error of the Mean.}

\item{iter}{Iterations in the variability analysis.}

\item{resolution}{Integer specifying the number of samples used in each hypercube dimension.}

\item{rng_init}{An integer value used to initialize the random number generator (RNG).}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Unmix sediment mixtures using constrained optimization and classical variability propagation
}
\keyword{internal}
