% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_string.R
\name{unmix_string}
\alias{unmix_string}
\title{Unmix sediment mixtures}
\usage{
unmix_string(data, samples = 100L, iter = 100L, Means = F, seed = 123456L)
}
\arguments{
\item{data}{Data frame containing sediment source and mixtures}

\item{samples}{Number of samples in each hypercube dimension}

\item{iter}{Iterations in the source variability analysis}

\item{Means}{Boolean to switch when using mean and sd data}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Asses the relative contribution of the potential sediment sources for each sediment mixture in the dataset.
}
