% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_unconstrained_lvp.R
\name{unmix_unconstrained_lvp}
\alias{unmix_unconstrained_lvp}
\title{Unmix sediment mixtures using unconstrained (least-squares) optimization and linear variability propagation (LVP)}
\usage{
unmix_unconstrained_lvp(
  data,
  variability = "SEM",
  iter = 1000,
  means = F,
  rng_init = 123456L
)
}
\arguments{
\item{data}{Data frame containing sediment source and mixtures}

\item{variability}{Character string specifying the type of variability to calculate.
Possible values are "SD" for Standard Deviation or "SEM" for Standard Error of the Mean.}

\item{iter}{Iterations in the source variability analysis.}

\item{means}{Boolean to switch when using mean and sd data}

\item{rng_init}{An integer value used to initialize the random number generator (RNG). Providing a starting value ensures that the sequence of random numbers generated is reproducible. This is useful for debugging, testing, and comparing results across different runs. If no value is provided, a random one will be generated.}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Asses the relative contribution of the potential sediment sources for each sediment mixture in the dataset.
}
\keyword{internal}
