% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unmixing_corrected_gof1}
\alias{unmixing_corrected_gof1}
\title{Mixing model}
\usage{
unmixing_corrected_gof1(sources, samples, trials = 100L, iter = 100L,
  seed = 69512L)
}
\arguments{
\item{sources}{Data frame containing sediment sources data}

\item{samples}{Data frame containing sediment mixtures data}

\item{trials}{Number of samples in each hypercube dimension}

\item{iter}{Iterations in the source variability analysis}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Mixing model
}
