% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_results.R
\name{validate_results}
\alias{validate_results}
\title{Evaluate the mathematical consistency of a tracer selection}
\usage{
validate_results(selected_data, apportionments, error_threshold = 0.05)
}
\arguments{
\item{selected_data}{A data frame containing the characteristics of sediment sources and mixtures for the specific tracer selection to be evaluated.}

\item{apportionments}{A numeric vector containing the apportionment values (contributions) to be evaluated for each source, in the same order as they appear in the data.}

\item{error_threshold}{A numeric value (e.g., 0.05) representing the maximum acceptable 
normalized error. This value is used as a benchmark to categorize tracers as consistent 
or inconsistent in the diagnostic messages.}
}
\value{
A data frame containing the normalized error for each tracer.
}
\description{
This function assesses the mathematical consistency of a tracer selection for an apportionment result by computing the normalized error between the predicted and observed tracer concentrations in the virtual mixture. A low normalized error for all tracers indicates a consistent tracer selection. This function can be used to diagnose problems in the results of fingerprinting models.
}
\details{
The function calculates a normalized error for each tracer to assess the consistency of a given apportionment solution. The method involves first computing a "virtual mixture" by using the proposed apportionment values to perform a weighted average of the source tracer concentrations. The error for each tracer is then the difference between the tracer concentration in the real mixture and the virtual mixture. This error is normalized by the range of the tracer, which is estimated from the extremes of the sources' confidence intervals.

A low normalized error for all tracers (i.e., less than a predefined threshold like $0.05$) indicates a mathematically consistent tracer selection. If most tracers show low errors while a few have high errors, it suggests that those tracers may be non-conservative or less influential on the model's result. Conversely, high normalized errors in most tracers indicate mathematical inconsistency and can point to the existence of multiple partial solutions in the dataset.
}
\references{
Latorre, B., Lizaga, I., Gaspar, L., & Navas, A. (2021). A novel method for analysing
consistency and unravelling multiple solutions in sediment fingerprinting.
*Science of The Total Environment*, *789*, 147804.
}
