% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Save the results}
\usage{
write_results(data)
}
\arguments{
\item{data}{Data frame containing the relative contribution of the potential sediment sources for each sediment mixture in the dataset}
}
\description{
The function saves the results in the workspace file for all the sediment mixture samples and for each sediment mixture sample separately
}
