% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResults.R
\name{writeResults}
\alias{writeResults}
\title{Save the results}
\usage{
writeResults(data)
}
\arguments{
\item{data}{Data frame containing the relative contribution of the potential sediment sources for each sediment mixture in the dataset}
}
\description{
The function saves the results in the workspace file for all the sediment mixture samples and for each sediment mixture sample separately
}
