% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchment}
\alias{catchment}
\title{Land use and fingerprinting properties in a Mediterranean catchment}
\format{A data frame with 22 rows and 25 variables:
\describe{
  \item{id}{reference number id of each sample analysed}
  \item{Land_Use}{grouping variable, in this study refers to the different land uses in the catchment}
  \item{Cs.137, K.40, Bi.214, Ra.226, Th.232, U.238, Ba, Nb, Zr, Sr, Rb, Pb, Zn, Fe, Mn, Cr, V, Ti, Ca, K, Al, Si, Mg}{value of the tracer property for each sample}
}}
\usage{
catchment
}
\description{
A dataset containing the different tracer properties of the different land uses in a Mediterranean catchment and two mixture samples located at the output of the catchment.
The variables are as follows:
}
\keyword{datasets}
