% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFATest.R
\name{DFATest}
\alias{DFATest}
\title{Discriminant function analysis test}
\usage{
DFATest(data, niveau = 0.1)
}
\arguments{
\item{data}{Data frame containing source and mixtures}

\item{niveau}{level for the approximate F-test decision}
}
\value{
Data frame only containing the variables that pass the DFA test
}
\description{
Performs a stepwise forward variable selection using the Wilk's Lambda criterion.
}
