% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{least_squares_c}
\alias{least_squares_c}
\title{least_squares_c}
\usage{
least_squares_c(sources, samples, iter = 100L, seed = 69512L)
}
\arguments{
\item{sources}{Data frame containing sediment source samples}

\item{samples}{Data frame containing mixture samples}

\item{iter}{Iterations in the source variability analysis}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution solved by the least squares method
}
\description{
least_squares_c
}
