% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_test.R
\name{range_test}
\alias{range_test}
\title{Range test}
\usage{
range_test(data)
}
\arguments{
\item{data}{Data frame containing source and mixtures}
}
\value{
Data frame containing sediment sources and mixtures
}
\description{
Function that excludes the properties of the sediment mixture/s outside the minimum and maximum values in the sediment sources.
}
