% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeTest.R
\name{rangeTest}
\alias{rangeTest}
\title{Range test}
\usage{
rangeTest(data)
}
\arguments{
\item{data}{Data frame containing source and mixtures}
}
\value{
Data frame containing sediment sources and mixtures
}
\description{
Function that excludes the properties of the sediment mixture/s outside the minimum and maximum values in the sediment sources.
}
