% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ternary_diagram.R
\name{Ternary_diagram}
\alias{Ternary_diagram}
\title{Ternary diagrams with the CI function results}
\usage{
Ternary_diagram(data, tracers = 1:2, n_row = 1, n_col = 2, Means = T)
}
\arguments{
\item{data}{Data frame containing the results from the CI_Method function}

\item{tracers}{Number of tracers to display}

\item{n_row}{Number of rows in which display the tracers}

\item{n_col}{Number of columns in which display the tracers}

\item{Means}{Number of iteration for each tracer}
}
\value{
Ternary diagrams from the individual tracer predictions
}
\description{
Display the CI prediction for the selected number
}
