% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triangles_random_c}
\alias{triangles_random_c}
\title{triangles_random_c}
\usage{
triangles_random_c(sources, samples, tracer = 0L, iter = 100L, seed = 69512L)
}
\arguments{
\item{sources}{Data frame containing sediment source samples}

\item{samples}{Data frame containing mixture samples}

\item{tracer}{Tracer in which implement the function}

\item{iter}{Iterations in the source variability analysis}

\item{seed}{Seed for the random number generator}
}
\value{
List of data frames containing all the possible prediction for each tracer
}
\description{
triangles_random_c
}
