% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix.R
\name{unmix}
\alias{unmix}
\title{Unmix sediment mixtures}
\usage{
unmix(data, samples = 100L, iter = 100L, seed = 123456L)
}
\arguments{
\item{data}{Data frame containing sediment source and mixtures}

\item{samples}{Number of samples in each hypercube dimension}

\item{iter}{Iterations in the source variability analysis}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Asses the relative contribution of the potential sediment sources for each sediment mixture in the dataset.
}
