% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unmix_c}
\alias{unmix_c}
\title{unmix_c}
\usage{
unmix_c(sources, samples, trials = 100L, iter = 100L, seed = 69512L)
}
\arguments{
\item{sources}{Data frame containing sediment source samples}

\item{samples}{Data frame containing mixture samples}

\item{trials}{Number of samples in each hypercube dimension}

\item{iter}{Iterations in the source variability analysis}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
unmix_c
}
